+++ test/aliyun_oss/object_test.exs
@@ -19,4 +19,22 @@ defmodule Aliyun.Oss.Client.ObjectTest do
       end
     end
   end
+
+  describe "sign_post_policy/2" do
+    test "sign post policy" do
+      policy = %{
+        "conditions" => [
+          ["content-length-range", 0, 10485760],
+          %{"bucket" => "ahaha"},
+          %{"A" => "a"},
+          %{"key" => "ABC"}
+        ],
+        "expiration" => "2013-12-01T12:00:00Z"
+      }
+      assert Object.sign_post_policy(policy, @access_key_secret) == %{
+        policy: "eyJjb25kaXRpb25zIjpbWyJjb250ZW50LWxlbmd0aC1yYW5nZSIsMCwxMDQ4NTc2MF0seyJidWNrZXQiOiJhaGFoYSJ9LHsiQSI6ImEifSx7ImtleSI6IkFCQyJ9XSwiZXhwaXJhdGlvbiI6IjIwMTMtMTItMDFUMTI6MDA6MDBaIn0=",
+        signature: "W835KpLsL6k1/oo28RcsEflB6hw="
+      }
+    end
+  end
 end